+++ addons/yat/src/commands/builtin/Whereami.cs
@@ -3,11 +3,15 @@ namespace YAT.Commands
 	[Command("whereami", "Prints the current scene name and path.", "[b]Usage[/b]: whereami", "wai")]
 	public partial class Whereami : ICommand
 	{
-		public CommandResult Execute(YAT yat, params string[] args)
-		{
-			var scene = yat.GetTree().CurrentScene;
+		public YAT Yat { get; set; }
 
-			yat.Terminal.Print($"{scene.GetPath()} ({scene.SceneFilePath})");
+		public Whereami(YAT Yat) => this.Yat = Yat;
+
+		public CommandResult Execute(params string[] args)
+		{
+			var scene = Yat.GetTree().CurrentScene;
+
+			Yat.Terminal.Print($"{scene.GetPath()} ({scene.SceneFilePath})");
 
 			return CommandResult.Success;
 		}